Fixed search and layout issues
Your search bar was jumping down on tablet because there was no room
for it once the menu items in the nav were revealed. I had to shorten
the <ul>, and the expanded width of the search bar on tablet and
desktop, to make them both fit on the same line.

Your main content on tablet was being offset by 2 even though you had
it at 10 columns. In order to have an even margins on both sides, you
needed to offset it by 1. Here is the logic: 1 column left + 10 column
width, and a remainder of 1 column left on the right side.